+++ ros2_wrapping/lidar_conversions/src/OusterToLidarNode.cxx
@@ -58,7 +58,7 @@ void OusterToLidarNode::Callback(const Pcl2_msg& msg_received)
 
   // Fill the map of device_id if the device hasn't already been attributed one
   if (this->DeviceIdMap.count(cloudO.header.frame_id) == 0)
-    this->DeviceIdMap[cloudO.header.frame_id] = this->DeviceIdMap.size();
+    this->DeviceIdMap[cloudO.header.frame_id] = (uint8_t)(this->DeviceIdMap.size());
 
   // We compute the rotation duration : to do so, we need to ignore the first frame of the LiDAR, but it doesn't really matter as it is just 100ms ignored
   // However, we chose to ignore the first frame after having initialized estimation parameters (see below)
